+++ at_functional_test/test/pkam_utils.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:typed_data';
 
 import 'package:crypto/crypto.dart';
 import 'package:crypton/crypton.dart';
@@ -7,17 +8,18 @@ import 'at_demo_data.dart';
 
 String generatePKAMDigest(String atsign, String challenge) {
   var privateKey = pkamPrivateKeyMap[atsign];
-  privateKey = privateKey.trim();
+  privateKey = privateKey!.trim();
   var key = RSAPrivateKey.fromString(privateKey);
   challenge = challenge.trim();
-  var sign = key.createSHA256Signature(utf8.encode(challenge));
+  var sign = key.createSHA256Signature(Uint8List.fromList(utf8.encode(challenge)));
   return base64Encode(sign);
 }
 
 /// Returns the digest of the user.
+/// Code for cram 
 String getDigest(String atsign, String key) {
   var secret = cramKeyMap[atsign];
-  secret = secret.trim();
+  secret = secret!.trim();
   var challenge = key;
   challenge = challenge.trim();
   var combo = '$secret$challenge';