+++ at_secondary/at_secondary_server/lib/src/exception/global_exception_handler.dart
@@ -5,8 +5,11 @@ import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_spec/at_persistence_spec.dart';
 import 'package:at_secondary/src/server/at_secondary_impl.dart';
 import 'package:at_server_spec/at_server_spec.dart';
+import 'package:at_utils/at_logger.dart';
 
 /// GlobalExceptionHandler class is used to handle all the exceptions in the system.
+var logger = AtSignLogger('GlobalExceptionHandler');
+
 class GlobalExceptionHandler {
   static final GlobalExceptionHandler _singleton =
       GlobalExceptionHandler._internal();
@@ -21,41 +24,63 @@ class GlobalExceptionHandler {
   /// params: AtException, AtConnection
   Future<void> handle(Exception exception,
       {AtConnection? atConnection, Socket? clientSocket}) async {
-    if (exception is InvalidSyntaxException ||
-        exception is InvalidAtSignException ||
-        exception is BlockedConnectionException ||
+    if (exception is InvalidAtSignException ||
         exception is UnAuthenticatedException ||
         exception is BufferOverFlowException ||
-        exception is IllegalArgumentException) {
-      // If we get InvalidSyntaxException, InvalidAtSignException, InboundConnectionLimitException
-      // send error code
-      // Close all the related inbound/outbound connections
-      await _sendResponse(exception as AtException, atConnection);
-      _closeConnection(atConnection);
-    } else if (exception is DataStoreException ||
+        exception is IllegalArgumentException ||
         exception is ConnectionInvalidException) {
-      // In case of DataStoreException
-      // Retry for n number of times and Close connection.
+      // All of these are SEVERE
+      logger.severe(exception.toString());
+      await _sendResponseForException(exception, atConnection);
+      // TODO but do they necessarily need the connection to be closed?
       _closeConnection(atConnection);
+
+    } else if (exception is BlockedConnectionException ||
+        exception is InvalidSyntaxException) {
+      // This is normal behaviour, log as INFO
+      logger.info(exception.toString());
+      await _sendResponseForException(exception, atConnection);
+
+      // We're closing the connection because
+      //   BlockedConnectionException thrown when the "from" atsign is on the "do not allow" list
+      //   InvalidSyntaxException is thrown because invalid syntax is rude, so we're rude in return
+      _closeConnection(atConnection);
+
+    } else if (exception is DataStoreException) {
+      logger.severe(exception.toString());
+      // TODO should we keep the connection open rather than closing it?
+      await _sendResponseForException(exception, atConnection);
+      _closeConnection(atConnection);
+
     } else if (exception is InboundConnectionLimitException) {
+      // This is SEVERE and requires different handling so we use _handleInboundLimit
+      logger.severe(exception.toString());
       await _handleInboundLimit(exception, clientSocket!);
+
     } else if (exception is OutboundConnectionLimitException ||
         exception is LookupException ||
         exception is SecondaryNotFoundException ||
         exception is HandShakeException ||
         exception is UnAuthorizedException ||
         exception is OutBoundConnectionInvalidException ||
-        exception is KeyNotFoundException) {
-      // In case of OutboundConnectionLimitException, LookupException, ConnectionInvalidException
-      // SecondaryNotFoundException, HandShakeException, UnAuthorizedException, UnverifiedConnectionException
-      // send error code.
-      await _sendResponse(exception as AtException, atConnection);
+        exception is KeyNotFoundException ||
+        exception is AtConnectException ||
+        exception is AtTimeoutException ||
+        exception is InvalidAtSignException ||
+        exception is UnAuthenticatedException) {
+      // TODO Not sure some of these are really worthy of WARNINGS, but let's leave as is for now
+      logger.warning(exception.toString());
+      await _sendResponseForException(exception, atConnection);
+
     } else if (exception is AtServerException ||
         exception is ArgParserException) {
       // In case of AtServerException terminate the server
+      logger.shout("Terminating secondary due to ${exception.toString()}");
       _terminateSecondary();
+
     } else if (exception is InternalServerError) {
       await _handleInternalException(exception, atConnection);
+
     } else {
       await _handleInternalException(
           InternalServerException(exception.toString()), atConnection);
@@ -65,9 +90,9 @@ class GlobalExceptionHandler {
 
   Future<void> _handleInboundLimit(
       AtException exception, Socket clientSocket) async {
-    var error_code = getErrorCode(exception);
-    var error_description = getErrorDescription(error_code);
-    clientSocket.write('error:$error_code-$error_description\n');
+    var errorCode = getErrorCode(exception);
+    var errorDescription = getErrorDescription(errorCode);
+    clientSocket.write('error:$errorCode-$errorDescription\n');
     await clientSocket.close();
   }
 
@@ -80,21 +105,27 @@ class GlobalExceptionHandler {
 
   Future<void> _handleInternalException(
       AtException exception, AtConnection? atConnection) async {
-    await _sendResponse(exception, atConnection);
+    await _sendResponseForException(exception, atConnection);
   }
 
   /// Method to write response to client
   /// Params: AtException, AtConnection
   /// We'll get error code based on the exception and write error:<error_code> to the client socket
-  Future<void> _sendResponse(
-      AtException exception, AtConnection? atConnection) async {
+  Future<void> _sendResponseForException(
+      Exception exception, AtConnection? atConnection) async {
     if (atConnection != null) {
       if (!atConnection.isInValid()) {
         var prompt = _getPrompt(atConnection);
-        var error_code = getErrorCode(exception);
-        var error_description =
-            '${getErrorDescription(error_code)} : ${exception.message}';
-        _writeToSocket(atConnection, prompt, error_code, error_description);
+        var errorCode = getErrorCode(exception);
+        errorCode ??= 'AT0011';
+
+        var errorDescription;
+        if (exception is AtException) {
+          errorDescription = '${getErrorDescription(errorCode)} : ${exception.message}';
+        } else {
+          errorDescription = '${getErrorDescription(errorCode)} : ${exception.toString()}';
+        }
+        _writeToSocket(atConnection, prompt, errorCode, errorDescription);
       }
     }
   }
@@ -111,16 +142,15 @@ class GlobalExceptionHandler {
   }
 
   String? getErrorCode(Exception exception) {
-    var error_code = error_codes[exception.runtimeType.toString()];
-    return error_code;
+    return error_codes[exception.runtimeType.toString()];
   }
 
-  String? getErrorDescription(String? error_code) {
-    return error_description[error_code];
+  String? getErrorDescription(String? errorCode) {
+    return error_description[errorCode];
   }
 
   void _writeToSocket(AtConnection atConnection, String prompt,
-      String? error_code, String error_description) {
-    atConnection.write('error:$error_code-$error_description\n$prompt');
+      String? errorCode, String errorDescription) {
+    atConnection.write('error:$errorCode-$errorDescription\n$prompt');
   }
 }