+++ at_secondary/at_persistence_secondary_server/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.16
+- at_lookup version upgrade for implementing server error responses
+- at_commons version upgrade for AtTimeoutException
+## 3.0.15
+- at_utils version upgrade
 ## 3.0.14
 - Fix commit log compaction issue.
 ## 3.0.13